+++ app/views/shared/_badges.html.erb
@@ -1,5 +1,5 @@
 
-<div class="badge__wrapper">
+<div class="badge__wrapper" data-badge-target="badgewrapper">
   <div class="badge__3d">
     <div class="badge__image">
       <div class="badge__layer1"></div>
