+++ app/views/shared/_navbar.html.erb
@@ -14,20 +14,19 @@
           <li class="nav-item name">
           <%if current_user.organisme == false %>
             <em><%= link_to current_user.firstname, dashboard_path , class: "nav-link" %></em>
-            <% else %>
+          <% else %>
             <em><%= link_to current_user.firstname, organisation_missions_path , class: "nav-link" %></em>
-            <% end %>
+          <% end %>
           </li>
           <li class="nav-item active">
             <%= link_to "Home", root_path, class: "nav-link" %>
           </li>
           <li class="nav-item dropdown">
-
-      <% if current_user.photo.key? %>
-        <%= cl_image_tag current_user.photo.key, class: "avatar dropdown-toggle", id: "navbarDropdown", data: { bs_toggle: "dropdown" }, 'aria-haspopup': true, 'aria-expanded': false %>
-      <% else %>
-            <%= image_tag "avatarFemme.jpg", class: "avatar dropdown-toggle", id: "navbarDropdown", data: { bs_toggle: "dropdown" }, 'aria-haspopup': true, 'aria-expanded': false %>
-      <% end %>
+            <% if current_user.photo.key? %>
+              <%= cl_image_tag current_user.photo.key, class: "avatar dropdown-toggle", id: "navbarDropdown", data: { bs_toggle: "dropdown" }, 'aria-haspopup': true, 'aria-expanded': false %>
+            <% else %>
+              <%= image_tag "avatarFemme.jpg", class: "avatar dropdown-toggle", id: "navbarDropdown", data: { bs_toggle: "dropdown" }, 'aria-haspopup': true, 'aria-expanded': false %>
+            <% end %>
 
             <div class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
               <% if current_user.organisme%>
